requirements.txt
cycler==0.10.0
h5py==2.8.0
kiwisolver==1.0.1
matplotlib==3.0.1
nltk==3.3
numpy==1.15.3
Pillow==5.3.0
pycocotools==2.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
PyYAML==3.13
six==1.11.0
torch==0.4.1
torchvision==0.2.1
